refactor: code review fixes across crates
- Remove dead kitty protocol code from terminal.rs
- Rename PI_ env var prefix to RHO_ in client
- Delete unused settings_list and stdin_buffer modules
- Consolidate duplicate validate_api_key tests into providers/mod.rs
- Extract shared FakeProvider/PendingProvider into rho-agent test_helpers
- Add Key::CTRL_C/CTRL_T constants to avoid hot-path allocations
- Add EditorState::is_multiline() helper
- Split 1570-line client.rs into client/{mod,app,render,input,network}.rs
- Replace unreachable post-loop error with unreachable!() in runtime
- Use unsigned_abs() instead of raw cast in overlay offset
