+++ crates/rho-agent/src/lib.rs
@@ -1,5 +1,8 @@
 pub mod runtime;
 pub mod server;
 
+#[cfg(test)]
+mod test_helpers;
+
 pub use runtime::{AgentError, AgentRuntime, AgentSession};
 pub use server::{AgentServer, AgentServerError, build_provider};
