+++ crates/rho-tui/src/lib.rs
@@ -4,8 +4,6 @@ mod editor;
 mod keys;
 mod overlay;
 mod select_list;
-mod settings_list;
-mod stdin_buffer;
 mod terminal;
 mod terminal_image;
 mod theme;
